♻️ refactor: add container layer to GridView for data and state
- introduced container component to manage data and state for GridView
- simplified GridView to focus solely on presentation
